requirements.txt
pyserial
streamlit
pandas
numpy